@@ -1,19 +1,9 @@
-export const ADD_FLASH_MESSAGE = 'ADD_FLASH_MESSAGE';
-export const DELETE_FLASH_MESSAGE = 'DELETE_FLASH_MESSAGE';
-
-export const CREATE_USER_SUCCESS = 'CREATE_USER_SUCCESS';
 export const SET_CURRENT_USER = 'SET_CURRENT_USER';
 export const RETRIEVE_USERS_SUCCESS = 'RETRIEVE_USERS_SUCCESS';
-export const RETRIEVE_USER_SUCCESSS = 'GET_USER_SUCCESSS';
-export const GET_USER_INFO_SUCCESS = 'GET_USER_INFO_SUCCESS';
-export const UPDATE_USER_SUCCESS = 'UPDATE_USER_SUCCESS';
-export const DELETE_USER = 'DELETE_USER';
 
-export const CREATE_DOCUMENT_SUCCESS = 'CREATE_DOCUMENT_SUCCESS';
-export const LOAD_DOCUMENT_SUCCESS = 'LOAD_DOCUMENT_SUCCESS';
-export const UPDATE_DOCUMENT_SUCCESS = 'UPDATE_DOCUMENT_SUCCESS';
-export const DELETE_CURRENT_DOCUMENT = 'DELETE_CURRENT_DOCUMENT';
-export const CHOOSE_AS_CURRENT_DOCUMENT = 'CHOOSE_AS_CURRENT_DOCUMENT';
+export const LOAD_ALL_DOCUMENTS = 'LOAD_ALL_DOCUMENTS';
+export const LOAD_USER_DOCUMENTS = 'LOAD_USER_DOCUMENTS';
 
-export const LOAD_ROLE_SUCCESS = 'LOAD_ROLE_SUCCESS';
+export const LOAD_ROLES_SUCCESS = 'LOAD_ROLES_SUCCESS';
 
+export const RESET = 'RESET';
