@@ -1,52 +1,45 @@
 import axios from 'axios';
-import jwtDecode from 'jwt-decode';
 import setAuthorizationToken from '../utils/setAuthorizationToken';
-import { SET_CURRENT_USER } from './actionTypes';
+import { SET_CURRENT_USER, RESET } from './actionTypes';
+import { loadUserDocuments } from './documentActions';
 
+const apiUrlPrefix = '/api';
 
-/**
- * set current user action creator
- * @export
- * @param {object} user
- * @returns {object} payload containing type and user object
- */
 export function setCurrentUser(user) {
   return {
     type: SET_CURRENT_USER,
-    user
+    payload: {
+      user
+    }
   };
 }
 
-
-/**
- * logout action creator
- * @export
- * @returns {function}
- */
 export function logout() {
-  return (dispatch) => {
+  return dispatch => {
     localStorage.removeItem('jwtToken');
     localStorage.removeItem('user');
     setAuthorizationToken(false);
-    dispatch(setCurrentUser({}));
+    dispatch({ type: RESET });
   };
 }
 
-
-/**
- * login action creator
- * @export
- * @param {object} data
- * @returns {function}
- */
 export function login(data) {
-  return dispatch => axios.post('/user/login', data)
-    .then((response) => {
-      const token = response.data.token;
-      const user = response.data.user;
-      localStorage.setItem('user', JSON.stringify(user));
-      localStorage.setItem('jwtToken', token);
-      setAuthorizationToken(token);
-      dispatch(setCurrentUser(jwtDecode(token)));
-    });
+  return async dispatch => {
+    const response = await axios.post(`${apiUrlPrefix}/user/login`, data);
+    const {
+      token,
+      username,
+      email,
+      firstname,
+      lastname,
+      roleId,
+      id
+    } = response.data;
+    const user = { username, email, firstname, lastname, roleId, id };
+    localStorage.setItem('user', JSON.stringify(user));
+    localStorage.setItem('jwtToken', token);
+    setAuthorizationToken(token);
+    dispatch(setCurrentUser(user));
+    dispatch(loadUserDocuments(id));
+  };
 }
