@@ -1,133 +1,75 @@
 import axios from 'axios';
-import * as types from './actionTypes';
+import { LOAD_ALL_DOCUMENTS, LOAD_USER_DOCUMENTS } from './actionTypes';
 
-/**
- * Load document success action creator
- * @export
- * @param {any} document
- * @returns {object} action
- */
-export function loadDocumentSuccess(document) {
-  return {
-    type: types.LOAD_DOCUMENT_SUCCESS,
-    document
-  };
+const apiUrlPrefix = '/api';
+
+export function retrieveAllDocuments(documents) {
+    return {
+        type: LOAD_ALL_DOCUMENTS,
+        payload: {
+            documents
+        }
+    };
 }
 
-
-/**
- * Update document success action creator
- * @export
- * @param {any} document
- * @returns {object} action
- */
-export function updateDocumentSuccess(document) {
-  return {
-    type: types.UPDATE_DOCUMENT_SUCCESS,
-    document
-  };
+export function retrieveUserDocuments(documents) {
+    return {
+        type: LOAD_USER_DOCUMENTS,
+        payload: {
+            documents
+        }
+    };
 }
 
-/**
- * Choose a document as the current document action creator
- * @export
- * @param {number} id
- * @returns {object} action
- */
-export function chooseAsCurrentDocument(id) {
-  return {
-    type: types.CHOOSE_AS_CURRENT_DOCUMENT,
-    id
-  };
-}
-
-/**
- * Delete the current document action creator
- * @return {object} actiontype
- */
-export function deleteCurrentDocument() {
-  return {
-    type: types.DELETE_CURRENT_DOCUMENT,
-  };
-}
-
-
-/**
- * action creator to get user documents
- * @param {number} user
- * @returns {function} documents
- */
 export function loadUserDocuments(id) {
-  return dispatch => axios.get(`user/${id}/document`)
-    .then((response) => {
-      dispatch(loadDocumentSuccess(response.data));
-    }).catch((error) => {
-      throw (error);
-    });
+    return async dispatch => {
+        axios.get(`${apiUrlPrefix}/user/${id}/document`).then(response => {
+            dispatch(retrieveUserDocuments(response.data));
+        });
+    };
 }
 
-/**
- * Action creator to get all documents accessible to current user
- * @returns {object} documents
- */
 export function loadAllDocuments() {
-  return dispatch => axios.get('/document')
-    .then((response) => {
-      dispatch(loadDocumentSuccess(response.data));
-    }).catch((error) => {
-      throw (error);
-    });
+    return async dispatch => {
+        axios.get(`${apiUrlPrefix}/document`).then(response => {
+            dispatch(retrieveAllDocuments(response.data));
+        });
+    };
 }
 
-/**
- * Action creator to save a document after adding content and title
- * @param {any} document
- * @param {any} userId
- * @returns {function}
- */
-export function saveDocument(document, id) {
-  return dispatch => axios.post('/document/', document)
-    .then(() => {
-      dispatch(loadUserDocuments(id));
-    }).catch((error) => {
-      throw (error);
-    });
-}
+export const saveDocument = (document, userId) => {
+    return async dispatch => {
+        axios.post(`${apiUrlPrefix}/document`, document).then(() => {
+            dispatch(loadUserDocuments(userId));
+            dispatch(loadAllDocuments());
+        });
+    };
+};
 
-/**
- * Action creator to update details of a document
- * @param {object} document
- * @returns {function}
- */
-export function updateDocument(document, userId) {
-  return dispatch => axios.put(`/document/${document.id}`, document)
-      .then(() => {
-        dispatch(loadUserDocuments(userId));
-      }).catch((error) => {
-        throw (error);
-      });
-}
+export const updateDocument = (document, userId) => {
+    return async dispatch => {
+        axios
+            .put(`${apiUrlPrefix}/document/${document.id}`, document)
+            .then(() => {
+                dispatch(loadUserDocuments(userId));
+            });
+    };
+};
 
-/**
- * @export
- * @param {any} id
- * @param {any} userId
- * @returns {object} documents
- */
 export function deleteDocument(id, userId) {
-  return dispatch => axios.delete(`/document/${id}`)
-    .then(() => {
-      dispatch(loadUserDocuments(userId));
-    }).catch((error) => {
-      throw (error);
-    });
+    return async dispatch => {
+        axios.delete(`${apiUrlPrefix}/document/${id}`).then(() => {
+            dispatch(loadUserDocuments(userId));
+        });
+    };
 }
 
-export function searchDocuments(query) {
-  return dispatch => axios.get(`/documents/search?query=${query}`)
-    .then((response) => {
-      dispatch(loadDocumentSuccess(response.data));
-    }).catch((error) => {
-      throw (error);
-    });
+export function searchForAllDocuments(query) {
+    return async dispatch => {
+        axios
+            .get(`${apiUrlPrefix}/documents/search?query=${query}`)
+            .then(response => {
+                dispatch(retrieveAllDocuments(response.data));
+            });
+    };
 }
