@@ -1,35 +1,38 @@
 import axios from 'axios';
-import jwtDecode from 'jwt-decode';
 import setAuthorizationToken from '../utils/setAuthorizationToken';
-import { SET_CURRENT_USER, CREATE_USER_SUCCESS } from './actionTypes';
+import { SET_CURRENT_USER } from './actionTypes';
+
+const apiUrlPrefix = '/api';
 
 export function setCurrentUser(user) {
-  return {
-    type: SET_CURRENT_USER,
-    user
-  };
+    return {
+        type: SET_CURRENT_USER,
+        payload: {
+            user
+        }
+    };
 }
 
-export function createUserSuccess(user) {
-  return {
-    type: CREATE_USER_SUCCESS,
-    user
-  };
+export function userAlreadyExists(identifier) {
+    return () => axios.get(`${apiUrlPrefix}/user/findUser/${identifier}`);
 }
 
-export function isUserExists(identifier) {
-  return (dispatch) => axios.get(`/user/findUser/${identifier}`);
+export function registerUser(userData) {
+    return async dispatch => {
+        const response = await axios.post(`${apiUrlPrefix}/user`, userData);
+        const {
+            username,
+            firstname,
+            lastname,
+            roleId,
+            email,
+            id,
+            token
+        } = response.data;
+        const user = { username, firstname, lastname, roleId, email, id };
+        localStorage.setItem('user', JSON.stringify(user));
+        localStorage.setItem('jwtToken', token);
+        setAuthorizationToken(token);
+        dispatch(setCurrentUser(user));
+    };
 }
-
-export function userSignupRequest(userData) {
-  return dispatch => axios.post('/user', userData)
-    .then((response) => {
-      const token = response.data.token;
-      const user = response.data.user;
-      localStorage.setItem('user', JSON.stringify(user));
-      localStorage.setItem('jwtToken', token);
-      setAuthorizationToken(token);
-      dispatch(setCurrentUser(jwtDecode(token)));
-    });
-}
-
