@@ -1,16 +1,16 @@
 import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import Header from './Common/Header';
-import FlashMessagesList from './FlashMessagesList/index';
+import { Header } from './base';
 import routes from '../routes';
+import CreateDocumentButton from './base/create-document-button';
 
 class App extends React.Component {
   render() {
     return (
       <BrowserRouter>
-        <React.Fragment>
-          <Header location={this.props.location}/>
+        <div className="app-root">
+          <Header />
           <Switch>
             {routes.map(({ component, exact, path }, index) => (
               <Route
@@ -21,15 +21,14 @@ class App extends React.Component {
               />
             ))}
           </Switch>
-          <FlashMessagesList />
-        </React.Fragment>
+          <CreateDocumentButton />
+        </div>
       </BrowserRouter>
     );
   }
 }
 
 App.propTypes = {
-  location: PropTypes.object,
   children: PropTypes.object
 };
 
