@@ -1,28 +1,33 @@
-/* eslint-disable import/no-extraneous-dependencies */
 import 'babel-polyfill';
 import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
-import jwtDecode from 'jwt-decode';
+import { persistStore } from 'redux-persist';
+import { PersistGate } from 'redux-persist/integration/react';
 import { setCurrentUser } from './actions/authenticationAction';
 import configureStore from './store/configureStore';
 import App from './components/App';
 import setAuthorizationToken from './utils/setAuthorizationToken';
-import '../node_modules/materialize-css/dist/js/materialize.min';
-import '../node_modules/materialize-css/dist/css/materialize.min.css';
-import '../node_modules/material-icons/css/material-icons.css';
-import '../node_modules/sweetalert/dist/sweetalert.css';
-import './styles/styles.css';
+import './styles/index.scss';
+import { retrieveUserFromToken } from './utils';
+import 'sweetalert2/dist/sweetalert2.min.css';
 
 const store = configureStore();
 
+const persistor = persistStore(store);
+
 if (localStorage.jwtToken) {
-  setAuthorizationToken(localStorage.jwtToken);
-  store.dispatch(setCurrentUser(jwtDecode(localStorage.jwtToken)));
+    setAuthorizationToken(localStorage.jwtToken);
+    retrieveUserFromToken(localStorage.jwtToken).then(decodedUser => {
+        store.dispatch(setCurrentUser(decodedUser));
+    });
 }
 
 render(
-  <Provider store={store}>
-    <App />
-  </Provider>, document.getElementById('app')
+    <Provider store={store}>
+        <PersistGate loading={null} persistor={persistor}>
+            <App />
+        </PersistGate>
+    </Provider>,
+    document.getElementById('app')
 );
