@@ -1,18 +1,14 @@
 import { combineReducers } from 'redux';
-import { reducer } from 'react-redux-sweetalert';
-import flashMessages from './flashMessagesReducer';
-import handleUsers from './userReducer';
-import handleDocuments from './documentReducer';
-import manageRoles from './roleReducer';
 import auth from './auth';
+import user from './user';
+import admin from './admin';
+import documents from './documents';
 
 export const rootReducer = combineReducers({
-  auth,
-  flashMessages,
-  handleUsers,
-  handleDocuments,
-  manageRoles,
-  sweetalert: reducer
+    auth,
+    admin,
+    user,
+    documents
 });
 
 export default rootReducer;
