@@ -1,58 +1,66 @@
-import HomePage from './components/HomePage';
-import AboutPage from './components/AboutPage';
-import SignupPage from './components/SignupPage';
-import LoginPage from './components/LoginPage';
-import DocumentPage from './components/DocumentPage';
-import DashboardPage from './components/DashboardPage';
-import ManangeRolePage from './components/adminComponents/ManageRolePage';
-import HandleUsersPage from './components/adminComponents/HandleUsersPage';
-import ProfilePage from './components/ProfilePage';
-import requireAuth from './components/Utils/RequireAuth';
-import requireAdminAuth from './components/Utils/RequireAdminAuth';
+import Home from './components/home';
+import About from './components/about';
+import Signup from './components/signup';
+import Login from './components/login';
+import { RenderDocument, Documents } from './components/document';
+import Dashboard from './components/dashboard';
+import { Users, Roles } from './components/adminComponents';
+import Profile from './components/profile';
+import { RequireAuth, RequireAdminAuth } from './components/protected';
 
 const routes = [
-  {
-    path: '/',
-    component: HomePage,
-    exact: true,
-  },
-  {
-    path: '/dashboard',
-    component: requireAuth(DashboardPage),
-    exact: true,
-  },
-  {
-    path: '/login',
-    component: LoginPage,
-    exact: true
-  },
-  {
-    path: '/signup',
-    component: SignupPage,
-    exact: true
-  },
-  {
-    path: '/documents',
-    component: requireAuth(DocumentPage),
-    exact: true
-  },
-  {
-    path: '/profilepage',
-    component: requireAuth(ProfilePage),
-    exact: true
-  },
-  {
-    path: '/about',
-    component: AboutPage
-  },
-  {
-    path: '/admin/manageroles',
-    component: requireAdminAuth(ManangeRolePage)
-  },
-  {
-    path: '/admin/handleusers',
-    component: requireAdminAuth(HandleUsersPage)
-  },
+    {
+        path: '/',
+        component: Home,
+        exact: true
+    },
+    {
+        path: '/dashboard',
+        component: RequireAuth(Dashboard),
+        exact: true
+    },
+    {
+        path: '/login',
+        component: Login,
+        exact: true
+    },
+    {
+        path: '/signup',
+        component: Signup,
+        exact: true
+    },
+    {
+        path: '/documents',
+        component: RequireAuth(Documents),
+        exact: true
+    },
+    {
+        path: '/document',
+        component: RequireAuth(RenderDocument),
+        exact: true
+    },
+    {
+        path: '/document/:id',
+        component: RequireAuth(RenderDocument),
+        exact: true
+    },
+    {
+        path: '/profile',
+        component: RequireAuth(Profile),
+        exact: true
+    },
+    {
+        path: '/about',
+        component: About
+    },
+    {
+        path: '/roles',
+        component: RequireAdminAuth(Roles)
+    },
+    {
+        path: '/users',
+        component: RequireAdminAuth(Users)
+    }
 ];
 
 export default routes;
