@@ -1,54 +1,39 @@
 module.exports = (sequelize, DataTypes) => {
-  const Document = sequelize.define('Document', {
-    title: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      validate: {
-        notEmpty: {
-          msg: 'You must provide a Title'
+    const Document = sequelize.define(
+        'Document',
+        {
+            title: {
+                type: DataTypes.STRING,
+                allowNull: false
+            },
+
+            content: {
+                type: DataTypes.TEXT,
+                allowNull: false
+            },
+
+            access: {
+                type: DataTypes.STRING,
+                defaultValue: 'public',
+                allowNull: false
+            },
+
+            ownerId: {
+                type: DataTypes.INTEGER,
+                allowNull: false
+            }
+        },
+        {
+            freezeTableName: true
         }
-      }
-    },
+    );
 
-    content: {
-      type: DataTypes.TEXT,
-      allowNull: false,
-      validate: {
-        notEmpty: {
-          msg: 'You cannot have an empty Document'
-        }
-      }
-    },
-
-    access: {
-      type: DataTypes.STRING,
-      defaultValue: 'public',
-      allowNull: false,
-      validate: {
-        isIn: [['private', 'public', 'role']]
-      }
-    },
-
-    ownerId: {
-      type: DataTypes.INTEGER,
-      allowNull: false,
-      validate: {
-        isInt: true
-      }
-    }
-
-  }, {
-    classMethods: {
-      associate(models) {
-          // associations can be defined here
+    Document.associate = models => {
         Document.belongsTo(models.User, {
-          as: 'owner',
-          onDelete: 'CASCADE',
-          foreignKey: { allowNull: true }
+            onDelete: 'CASCADE',
+            foreignKey: 'ownerId'
         });
-      }
-    },
-    freezeTableName: true
-  });
-  return Document;
+    };
+
+    return Document;
 };
