@@ -1,77 +1,73 @@
-import bcrypt from 'bcrypt-nodejs';
+import bcrypt from 'bcrypt';
 
 module.exports = (sequelize, DataTypes) => {
-  const User = sequelize.define('User', {
-    userName: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      unique: true
-    },
-    firstName: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    lastName: {
-      type: DataTypes.STRING,
-      allowNull: false
-    },
-    email: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      unique: true,
-      validate: { isEmail: true }
-    },
-    password: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      validate: { len: 6 }
-    },
-    roleId: {
-      type: DataTypes.INTEGER,
-      allowNull: false,
-      defaultValue: 2
-    }
-  }, {
-    classMethods: {
-      associate: (models) => {
-        // associations can be defined here
+    const User = sequelize.define(
+        'User',
+        {
+            username: {
+                type: DataTypes.STRING,
+                allowNull: false,
+                unique: true
+            },
+            firstname: {
+                type: DataTypes.STRING,
+                allowNull: false
+            },
+            lastname: {
+                type: DataTypes.STRING,
+                allowNull: false
+            },
+            email: {
+                type: DataTypes.STRING,
+                allowNull: false,
+                unique: true,
+                validate: { isEmail: true }
+            },
+            password: {
+                type: DataTypes.STRING,
+                allowNull: false,
+                validate: { len: 6 }
+            },
+            roleId: {
+                type: DataTypes.INTEGER,
+                allowNull: false,
+                defaultValue: 2
+            }
+        },
+        {
+            freezeTableName: true,
+
+            hooks: {
+                beforeCreate: user => user.password && user.hashPassword(),
+                beforeUpdate: user => {
+                    if (user.changed('password')) {
+                        return user.hashPassword();
+                    }
+                }
+            }
+        }
+    );
+
+    User.associate = models => {
         User.belongsTo(models.Role, {
-          onDelete: 'CASCADE',
-          foreignKey: 'roleId'
+            onDelete: 'CASCADE',
+            foreignKey: 'roleId'
         });
+
         User.hasMany(models.Document, {
-          onDelete: 'CASCADE',
-          foreignKey: 'ownerId'
+            onDelete: 'CASCADE',
+            foreignKey: 'ownerId'
         });
-      }
-    },
-    freezeTableName: true,
+    };
 
-    instanceMethods: {
-      /**
-       * Compare plain password to hashed one
-       * @method
-       * @param {String} password
-       * @returns {Boolean} password match
-       */
-      validPassword(password) {
-        return bcrypt.compareSync(password, this.password);
-      },
+    User.prototype.hashPassword = async function hashPassword() {
+        this.password = await bcrypt.hash(this.password, bcrypt.genSaltSync(9));
+        return this.password;
+    };
 
-      /**
-       * Hash password received from user
-       * @method
-       * @returns {void}
-       */
-      hashPassword() {
-        this.password = bcrypt.hashSync(this.password, bcrypt.genSaltSync(9));
-      }
-    },
-    hooks: {
-      beforeCreate(user) {
-        user.hashPassword();
-      }
-    }
-  });
-  return User;
+    User.prototype.validPassword = async function validPassword(password) {
+        const match = await bcrypt.compare(password, this.password);
+        return match;
+    };
+    return User;
 };
