@@ -2,38 +2,39 @@ import Validator from 'validator';
 import isEmpty from 'lodash/isEmpty';
 
 export default function validateInput(data) {
-  const errors = {};
+    console.log('Supplied Data', data);
+    const errors = {};
 
-  if (Validator.isEmpty(data.userName)) {
-    errors.userName = 'Username is required';
-  }
-  if (Validator.isEmpty(data.firstName)) {
-    errors.firstName = 'First Name is required';
-  }
-  if (Validator.isEmpty(data.lastName)) {
-    errors.lastName = 'Last Name is required';
-  }
-  if (Validator.isEmpty(data.email)) {
-    errors.email = 'Email is invalid';
-  }
-  if (!Validator.isEmail(data.email)) {
-    errors.email = 'Email is invalid';
-  }
-  if (Validator.isEmpty(data.password)) {
-    errors.password = 'Password is required';
-  }
-  if (!Validator.isLength(data.password, { min: 6 })) {
-    errors.password = 'Password must be at least 6 characters';
-  }
-  if (Validator.isEmpty(data.passwordConfirmation)) {
-    errors.passwordConfirmation = 'Passwords must match';
-  }
-  if (!Validator.equals(data.password, data.passwordConfirmation)) {
-    errors.passwordConfirmation = 'Passwords must match';
-  }
+    if (Validator.isEmpty(data.username)) {
+        errors.userName = 'Username is required';
+    }
+    if (Validator.isEmpty(data.firstname)) {
+        errors.firstName = 'First Name is required';
+    }
+    if (Validator.isEmpty(data.lastname)) {
+        errors.lastName = 'Last Name is required';
+    }
+    if (Validator.isEmpty(data.email)) {
+        errors.email = 'Email is required';
+    }
+    if (!Validator.isEmail(data.email)) {
+        errors.email = 'Email is invalid';
+    }
+    if (Validator.isEmpty(data.password)) {
+        errors.password = 'Password is required';
+    }
+    if (!Validator.isLength(data.password, { min: 6 })) {
+        errors.password = 'Password must be at least 6 characters';
+    }
+    if (Validator.isEmpty(data.confirmPassword)) {
+        errors.confirmPassword = 'Passwords must match';
+    }
+    if (!Validator.equals(data.password, data.confirmPassword)) {
+        errors.confirmPassword = 'Passwords must match';
+    }
 
-  return {
-    errors,
-    isValid: isEmpty(errors)
-  };
+    return {
+        errors,
+        isValid: isEmpty(errors)
+    };
 }
